Lab01/requirements.txt
streamlit
azure-storage-blob
pymssql==2.3.1
dotenv